Struct fftw3::builder::Planner

pub struct Planner {
    // some fields omitted
}

Control the basic properties of a set of transforms.

Methods

impl Planner

fn new() -> Planner

Construct a new planner with default values.

This defaults to a forward transform with estimate rigor.

fn rigor(self, r: Rigor) -> Planner

Set the rigor to use for this plan.

fn wisdom_restriction(self, wisdom_only: bool) -> Planner

Set whether the planner should only be successfully created if there exists wisdom created with at least the rigor level set.

fn direction(self, direction: Direction) -> Planner

Set the direction of the transform to perform.

fn inplace(self) -> InPlacePlanner

fn c2c<I, O>(self, in_: I, out: O) -> PlanMem<I, O> where I: DerefMut<[Complex64]>, O: DerefMut<[Complex64]>

fn c2r<I, O>(self, in_: I, out: O) -> PlanMem<I, O> where I: DerefMut<[Complex64]>, O: DerefMut<[f64]>

fn r2c<I, O>(self, in_: I, out: O) -> PlanMem<I, O> where I: DerefMut<[f64]>, O: DerefMut<[Complex64]>