Struct fftw3::builder2::Planner

pub struct Planner<Stage> {
    // some fields omitted
}

Methods

impl Planner<Begin>

fn new() -> Planner<Begin>

fn input<T, I: MutStrided<T>>(self, in_: I) -> Planner<Input<I>>

impl Planner<Begin>

fn rigor(self, r: Rigor) -> Planner<Begin>

Set the rigor to use for this plan.

fn wisdom_restriction(self, wisdom_only: bool) -> Planner<Begin>

Set whether the planner should only be successfully created if there exists wisdom created with at least the rigor level set.

fn direction(self, direction: Direction) -> Planner<Begin>

Set the direction of the transform to perform.

fn _1d(self, n: uint) -> Planner<Begin>

fn _2d(self, n0: uint, n1: uint) -> Planner<Begin>

fn _3d(self, n0: uint, n1: uint, n2: uint) -> Planner<Begin>

fn nd(self, dims: &[uint]) -> Planner<Begin>

impl<T, I: MutStrided<T>> Planner<Input<I>>

fn output<U: FftData<T>, O: MutStrided<U>>(self, out: O) -> Planner<Io<I, O>>

impl<T: FftData<T>, I: MutStrided<T>> Planner<Input<I>>

fn inplace(self) -> Planner<Inplace<I>>

impl<X: FftSpec> Planner<X>

fn plan(self) -> Result<Plan<X>, PlanningError>