Struct fftw3::wisdom::WisdomGuard

pub struct WisdomGuard {
    // some fields omitted
}

Import and export FFTW wisdom implicitly.

The destructor will save wisdom to the file from which it was loaded.

The wisdom macro performs this automatically.

Example

{
   wisdom!(Path::new("./wise.fftw"));
   // ... perform FFTs with the assistance of that wisdom

} // any new wisdom is automatically saved.

Methods

impl WisdomGuard

fn import(p: Path) -> WisdomGuard

Load wisdom from p, and save it automatically on clean-up.

Failure to load is ignored, so one can supply a wisdom file that does not exist (yet) and it will be created on the next run.